pyproject.toml
[tool.poetry]
name = "eticketing_api"
version = "0.1.0"
description = "Eticketing API"
authors = ["mogendi <mogendi.mongare@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.3.2"
black = "21.11b1"
mypy = "^0.931"
isort = "^5.8.0"
autoflake = "^1.4"
Pillow = "^9.0.0"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
pydantic = "^1.9.0"
drf-yasg = "^1.20.0"
flake8 = ">=3.7.9"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyYAML = "^6.0"

[tool.isort]
profile = "black"
[tool.black]
target-version = ['py37', 'py38', 'py39']
exclude = '''
(
    __pycache__
  | .pytest_cache
  | .gitignore
  | ^*.html
)
'''

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
